requirements.txt
aenum==3.1.15
annotated-types==0.6.0
anthropic==0.26.1
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
django-cors-headers==4.3.1
elastic-transport==8.12.0
elasticsearch==8.12.0
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.2
Flask-Cors==4.0.1
fqdn==1.5.1
fsspec==2024.5.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
gradientai==1.12.1
grpcio==1.64.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.23.1
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jiter==0.1.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.11
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
langchain-core==0.1.16
langchain-openai==0.0.5
langsmith==0.0.83
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
openai==1.10.0
openpyxl==3.1.2
overrides==7.6.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==1.10.15
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.1
regex==2023.12.25
replicate==0.24.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
tenacity==8.2.3
terminado==0.18.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.19.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
views==0.3
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
